# server/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: 若依 Go 版本接口文档
  title: RuoYi-Go API
  version: "1.0"
paths:
...
  /monitor/job:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 新增定时任务
      tags:
      - 定时任务
    put:
      consumes:
      - application/json
      parameters:
      - description: 任务信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改定时任务
      tags:
      - 定时任务
  /monitor/job/{id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 任务信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询定时任务详细
      tags:
      - 定时任务
  /monitor/job/{ids}:
    delete:
      parameters:
      - description: 任务ID，多个用逗号分隔
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除定时任务
      tags:
      - 定时任务
  /monitor/job/changeStatus:
    put:
      consumes:
      - application/json
      parameters:
      - description: 包含id和status
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改任务状态
      tags:
      - 定时任务
  /monitor/job/export:
    post:
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Excel文件
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出定时任务数据
      tags:
      - 定时任务
  /monitor/job/list:
    get:
      parameters:
      - description: 任务名称
        in: query
        name: jobName
        type: string
      - description: 任务组名
        in: query
        name: jobGroup
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 任务列表
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询定时任务列表
      tags:
      - 定时任务
  /monitor/job/run:
    put:
      responses: {}
      security:
      - BearerAuth: []
      summary: 立即执行一次任务
      tags:
      - 定时任务
...
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"